modularized face_detection nms code
