+++ src/mediapipe_solutions/mod.rs
@@ -1,0 +1,3 @@
+pub mod face_detection;
+
+// use crate::mediapipe_solutions::face_detection;
