+++ src/show_output/mod.rs
@@ -1,0 +1,2 @@
+pub mod print_output;
+pub mod draw_output;
